modified:   Makefile 	modified:   src/Bullet.cpp 	modified:   src/Bullet.h 	modified:   src/Character.cpp 	modified:   src/Character.h 	modified:   src/Fighter.cpp 	modified:   src/GCharacterSelectState.cpp 	modified:   src/GCharacterSelectState.h 	modified:   src/GMenuState.cpp 	modified:   src/GMenuState.h 	modified:   src/GPlayState.cpp 	modified:   src/GameState.h 	modified:   src/Ninja.cpp 	modified:   src/Shooter.cpp 	modified:   src/Shooter.h 	modified:   src/main.cpp

+++ src/Character.cpp
@@ -66,7 +66,7 @@ void Character::stop()
     state = STANDING;
 }
 
-void Character::damage(int hAmount, int sAmount, float speed, cpVect knockBack)
+void Character::damage(float hAmount, int sAmount, float speed, cpVect knockBack)
 {
     health -= hAmount;
     stamina -= sAmount;
